Support reading GitHub token from env var
I would like to share my config.yaml and update scripts in a git repo, so I need a way to supply the GitHub token outside of config.yaml

This solution felt natural to me, but I would be willing to take a second pass utilizing a credentials file instead.

Thanks for this great tool!